+++ test/borsh.cpp
@@ -4,7 +4,11 @@
 #include <vector>
 #include <string>
 
-#include "borsh.hpp"
+#include "concepts.h"
+#include "utils.h"
+#include "converters.h"
+#include "serializer.h"
+#include "borsh.h"
 #include "boost/ut.hpp"
 
 struct Vector2D
